backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
google-generativeai>=0.3.0
google-cloud-documentai>=2.20.0
google-cloud-dlp>=3.12.0
pymupdf>=1.23.0
pypdf>=3.0.0
python-docx>=0.8.11
requests>=2.28.0
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
aiofiles>=23.0.0
Pillow>=10.0.0
psycopg2-binary
